backend/requirements.txt
Flask
torch
torchvision
Pillow
numpy
pandas
scikit-learn
opencv-python
gunicorn
